MB-43550 Get specific stats in bulk from KVBucket, KVStore
Currently we gather kvstore stats for engine stats reporting by calling
KVStore individually for each stat. This is inefficient for MagmaKVStore
since for every such call, it fetches all its stats and only returns the
one asked for. It has to do this because Magma only supports getting all
stats at once and has no provision to get individual stats.

To help with such cases, this change adds a method to KVBucket and
KVStore to get specific stats in bulk.

Change-Id: I1c17f672bbf1ac45959354cb1e970cce65953f61